move gql-utils to utils.js
